Tidy up startup
