+++ src/web/EmployerFavouritesWeb/Startup.Configure.cs
@@ -1,76 +1,11 @@
-﻿using System.Linq;
-using DfE.EmployerFavourites.Web.Security;
-using DfE.EmployerFavourites.Web.Configuration;
-using MediatR;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Formatters;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using DfE.EmployerFavourites.Web.Infrastructure;
-using DfE.EmployerFavourites.Web.Domain;
-using SFA.DAS.EAS.Account.Api.Client;
-using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
 
 namespace DfE.EmployerFavourites.Web
 {
-    public class Startup
+    public partial class Startup
     {
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
-        {
-            Configuration = configuration;
-            _hostingEnvironment = env;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        private IHostingEnvironment _hostingEnvironment;
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMediatR(typeof(Startup).Assembly);
-            
-            services.Configure<CookiePolicyOptions>(options =>
-            {
-                // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-                options.CheckConsentNeeded = context => true;
-                options.MinimumSameSitePolicy = SameSiteMode.None;
-            });
-
-            services.AddMvc(options => {
-                // Add media type for csp report.
-                var jsonInputFormatters = options.InputFormatters.OfType<JsonInputFormatter>();
-                foreach (var formatter in jsonInputFormatters)
-                {
-                    formatter.SupportedMediaTypes
-                        .Add(Microsoft.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/csp-report"));
-                }
-            }).SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
-            JsonConvert.DefaultSettings = () =>
-            {
-                var settings = new JsonSerializerSettings();
-                settings.ContractResolver = new CamelCasePropertyNamesContractResolver();
-                return settings;
-            };
-
-            services.AddAuthenticationService(Configuration, _hostingEnvironment);
-
-            services.Configure<ExternalLinks>(Configuration.GetSection("ExternalLinks"));
-            services.Configure<ConnectionStrings>(Configuration.GetSection("ConnectionStrings"));
-
-            services.AddSingleton<IAccountApiConfiguration>(x => Configuration.GetSection("AccountApiConfiguration").Get<AccountApiConfiguration>());
-            services.AddTransient<IAccountApiClient, AccountApiClient>();
-
-            services.AddScoped<IFavouritesRepository, AzureTableStorageFavouritesRepository>(); 
-        }
-
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, IApplicationLifetime applicationLifetime, ILogger<Startup> logger)
         {
